Adding support for lines with multiple hostnames, and tests
The original version of this resource assumed that there would only be one hostname for each IP address, and would not function properly if this were not the case.  The code has been updated to handle this condition.

There was also a bug in the old version's regex pattern that would match partial IP addresses (eg, if a line contained 11.2.3.4 and the IPAddress parameter to the resource was 1.2.3.4, the resource would say that the IP was found.)  This bug has been fixed in this update.

Added Pester tests.